+++ src/js/shariff.js
@@ -152,10 +152,11 @@ class Shariff {
   }
 
   getTitle() {
-    let title = this.getOption('title') || this.getMeta('DC.title')
+    let title = this.getOption('title')
+    if ($(this.element).data()['title']) return title
+    title = title || this.getMeta('DC.title')
     let creator = this.getMeta('DC.creator')
-    if (title && creator) title = `${title} - ${creator}`
-    return title
+    return (title && creator) ? `${title} - ${creator}` : title
   }
 
   getReferrerTrack() {